+++ src/main/kotlin/omok/model/game/OmokRule.kt
@@ -1,6 +1,5 @@
-package omok.model.judge
+package omok.model.game
 
-import omok.model.Board
 import omok.model.state.State
 import omok.model.state.Stay
 import omok.model.state.Win
