+++ src/main/kotlin/omok/model/state/State.kt
@@ -1,5 +1,5 @@
 package omok.model.state
 
-import omok.model.Board
+import omok.model.game.Board
 
 abstract class State(val board: Board)
