+++ src/test/kotlin/omok/model/stone/CoordinateTest.kt
@@ -1,6 +1,5 @@
-package omok.model
+package omok.model.stone
 
-import omok.model.stone.Coordinate
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.Test
